packages/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Connor McLean <c.mclean165@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
SQLAlchemy = "^2.0.25"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
loguru = "^0.7.2"
passlib = "^1.7.4"
pydantic = "^2.6.1"
Authlib = "^1.3.0"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
PyJWT = "^2.8.0"
python-multipart = "^0.0.9"
itsdangerous = "^2.1.2"
stripe = "^8.2.0"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
redis = "^5.0.4"
boto3 = "^1.34.106"
twilio = "^9.0.5"
pynamodb = "^6.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
